apps/bridge_api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
aiohttp==3.9.1
python-dotenv==1.0.0
python-multipart==0.0.6
websockets==12.0